+++ lib/bolt/inventory.rb
@@ -234,7 +234,7 @@ module Bolt
       set_facts(target.name, data['facts']) unless @target_facts[target.name]
       data['features']&.each { |feature| set_feature(target, feature) } unless @target_features[target.name]
       unless @target_plugin_hooks[target.name]
-        set_plugin_hooks(target.name, @config.plugin_hooks.merge(data['plugin_hooks'] || {}))
+        set_plugin_hooks(target.name, (@plugins&.plugin_hooks || {}).merge(data['plugin_hooks'] || {}))
       end
 
       # Use Config object to ensure config section is treated consistently with config file