+++ lib/bolt/plugin.rb
@@ -39,9 +39,10 @@ module Bolt
     end
 
     class PluginContext
-      def initialize(config, pal)
+      def initialize(config, pal, plugins)
         @pal = pal
         @config = config
+        @plugins = plugins
       end
 
       def serial_executor
@@ -50,7 +51,7 @@ module Bolt
       private :serial_executor
 
       def empty_inventory
-        @empty_inventory ||= Bolt::Inventory.new({}, @config)
+        @empty_inventory ||= Bolt::Inventory::Inventory2.new({}, @config, plugins: @plugins)
       end
       private :empty_inventory
 
@@ -120,29 +121,46 @@ module Bolt
 
     def self.setup(config, pal, pdb_client, analytics)
       plugins = new(config, pal, analytics)
-      # PDB is special do we want to expose the default client to the context?
+
+      # PDB is special because it needs the PDB client. Since it has no config,
+      # we can just add it first.
       plugins.add_plugin(Bolt::Plugin::Puppetdb.new(pdb_client))
 
-      plugins.add_ruby_plugin('Bolt::Plugin::InstallAgent')
-      plugins.add_ruby_plugin('Bolt::Plugin::Task')
-      plugins.add_ruby_plugin('Bolt::Plugin::Pkcs7')
-      plugins.add_ruby_plugin('Bolt::Plugin::Prompt')
+      # Initialize any plugins referenced in config. This will also indirectly
+      # initialize any plugins they depend on.
+      if plugins.reference?(config.plugins)
+        msg = "The 'plugins' setting cannot be set by a plugin reference"
+        raise PluginError.new(msg, 'bolt/plugin-error')
+      end
+
+      config.plugins.keys.each do |plugin|
+        plugins.by_name(plugin)
+      end
+
+      plugins.plugin_hooks.merge!(plugins.resolve_references(config.plugin_hooks))
 
       plugins
     end
 
-    BUILTIN_PLUGINS = %w[task terraform pkcs7 prompt vault aws_inventory
-                         puppetdb azure_inventory yaml].freeze
+    RUBY_PLUGINS = %w[install_agent task pkcs7 prompt].freeze
+    BUILTIN_PLUGINS = %w[task terraform pkcs7 prompt vault aws_inventory puppetdb azure_inventory yaml].freeze
+    DEFAULT_PLUGIN_HOOKS = { 'puppet_library' => { 'plugin' => 'puppet_agent', 'stop_service' => true } }.freeze
 
     attr_reader :pal, :plugin_context
+    attr_accessor :plugin_hooks
+
+    private_class_method :new
 
     def initialize(config, pal, analytics)
       @config = config
       @analytics = analytics
-      @plugin_context = PluginContext.new(config, pal)
+      @plugin_context = PluginContext.new(config, pal, self)
       @plugins = {}
       @pal = pal
       @unknown = Set.new
+      @resolution_stack = []
+      @unresolved_plugin_configs = config.plugins.dup
+      @plugin_hooks = DEFAULT_PLUGIN_HOOKS.dup
     end
 
     def modules
@@ -154,11 +172,11 @@ module Bolt
       @plugins[plugin.name] = plugin
     end
 
-    def add_ruby_plugin(cls_name)
-      snake_name = Bolt::Util.class_name_to_file_name(cls_name)
-      require snake_name
-      cls = Kernel.const_get(cls_name)
-      plugin_name = snake_name.split('/').last
+    def add_ruby_plugin(plugin_name)
+      cls_name = Bolt::Util.snake_name_to_class_name(plugin_name)
+      filename = "bolt/plugin/#{plugin_name}"
+      require filename
+      cls = Kernel.const_get("Bolt::Plugin::#{cls_name}")
       opts = {
         context: @plugin_context,
         config: config_for_plugin(plugin_name)
@@ -178,14 +196,18 @@ module Bolt
       add_plugin(plugin)
     end
 
-    def add_from_config
-      @config.plugins.keys.each do |plugin_name|
-        by_name(plugin_name)
-      end
-    end
-
     def config_for_plugin(plugin_name)
-      @config.plugins[plugin_name] || {}
+      return {} unless @unresolved_plugin_configs.include?(plugin_name)
+      if @resolution_stack.include?(plugin_name)
+        msg = "Configuration for plugin '#{plugin_name}' depends on the plugin itself"
+        raise PluginError.new(msg, 'bolt/plugin-error')
+      else
+        @resolution_stack.push(plugin_name)
+        config = resolve_references(@unresolved_plugin_configs[plugin_name])
+        @unresolved_plugin_configs.delete(plugin_name)
+        @resolution_stack.pop
+        config
+      end
     end
 
     def get_hook(plugin_name, hook)
@@ -201,7 +223,9 @@ module Bolt
     def by_name(plugin_name)
       return @plugins[plugin_name] if @plugins.include?(plugin_name)
       begin
-        unless @unknown.include?(plugin_name)
+        if RUBY_PLUGINS.include?(plugin_name)
+          add_ruby_plugin(plugin_name)
+        elsif !@unknown.include?(plugin_name)
           add_module_plugin(plugin_name)
         end
       rescue PluginError::Unknown