+++ spec/bolt/plugin_spec.rb
@@ -7,8 +7,7 @@ require 'bolt/pal'
 require 'bolt/plugin'
 require 'bolt/analytics'
 
-# TODO: This is probably just for Bolt::Plugin::Module
-describe Bolt::Plugin::Module do
+describe Bolt::Plugin do
   include BoltSpec::Files
   include BoltSpec::Config
 
@@ -16,8 +15,16 @@ describe Bolt::Plugin::Module do
   let(:plugin_config) { {} }
   let(:config_data) { { 'modulepath' => modulepath, 'plugins' => plugin_config } }
   let(:pal) { Bolt::PAL.new(modulepath, nil, nil) }
+  let(:pdb_client) { double('pdb_client') }
 
-  let(:plugins) { Bolt::Plugin.new(config(config_data), pal, Bolt::Analytics::NoopClient.new) }
+  let(:plugins) { Bolt::Plugin.setup(config(config_data), pal, pdb_client, Bolt::Analytics::NoopClient.new) }
+
+  def identity(value)
+    {
+      '_plugin' => 'identity',
+      'value' => value
+    }
+  end
 
   it 'loads an empty plugin module' do
     expect(plugins.by_name('empty_plug').hooks).to eq([:validate_resolve_reference])
@@ -35,5 +42,63 @@ describe Bolt::Plugin::Module do
     expect(hooks).to include(:resolve_reference)
     expect(hooks).to include(:createkeys)
     expect(hooks).not_to include(:decrypt)
+  end
+
+  context 'evaluating plugin config' do
+    it 'lets a plugin depend on another plugin' do
+      plugin_config.replace('pkcs7' => { 'keysize' => identity(1024) })
+      expect(plugins.by_name('pkcs7').keysize).to eq(1024)
+    end
+
+    it 'fails if a plugin depends on itself' do
+      plugin_config.replace('identity' => { 'foo' => identity('bar') })
+      expect { plugins }.to raise_error(/Configuration for plugin 'identity' depends on the plugin itself/)
+    end
+
+    it 'fails if an indirect plugin dependency cycle is found' do
+      plugin_config.replace('pkcs7' => { 'keysize' => identity(1024) },
+                            'identity' => { 'foo' => { '_plugin' => 'pkcs7' } })
+      expect { plugins }.to raise_error(/Configuration for plugin 'pkcs7' depends on the plugin itself/)
+    end
+
+    it 'fails if the entire plugins key is set with a reference' do
+      plugin_config.replace(identity('pkcs7' => { 'keysize' => 1024 }))
+
+      expect { plugins }.to raise_error(/The 'plugins' setting cannot be set by a plugin reference/)
+    end
+  end
+
+  context 'loading plugin_hooks' do
+    it 'evaluates plugin references in the plugin_hooks configuration' do
+      config_data['plugin_hooks'] = {
+        'puppet_library' => {
+          'plugin' => 'my_hook',
+          'param' => identity('foobar')
+        }
+      }
+      expect(plugins.plugin_hooks['puppet_library']).to eq('plugin' => 'my_hook', 'param' => 'foobar')
+    end
+
+    it 'allows the whole plugin_hooks value to be set with a reference' do
+      hooks = {
+        'another_hook' => {
+          'plugin' => 'my_hook',
+          'param' => identity('foobar')
+        }
+      }
+
+      config_data['plugin_hooks'] = identity(hooks)
+
+      expect(plugins.plugin_hooks).to eq(
+        'another_hook' => {
+          'plugin' => 'my_hook',
+          'param' => 'foobar'
+        },
+        'puppet_library' => {
+          'plugin' => 'puppet_agent',
+          'stop_service' => true
+        }
+      )
+    end
   end
 end